1. Added Services as test resources to have code with ST in more readable form. 2. Created general methods fot ST 3. Splitted out ST for Consumer and Topics
