+++ src/test/java/io/strimzi/kafka/bridge/http/HttpBridgeTest.java
@@ -23,14 +23,9 @@ import io.vertx.ext.web.client.WebClientOptions;
 import io.vertx.ext.web.codec.BodyCodec;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
-import io.vertx.kafka.client.consumer.KafkaConsumer;
-import io.vertx.kafka.client.consumer.KafkaConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.consumer.OffsetResetStrategy;
 import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.common.serialization.ByteArrayDeserializer;
 import org.apache.kafka.common.serialization.ByteArraySerializer;
-import org.apache.kafka.common.serialization.StringDeserializer;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
@@ -43,7 +38,6 @@ import javax.xml.bind.DatatypeConverter;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
@@ -71,27 +65,24 @@ class HttpBridgeTest extends KafkaClusterTestBase {
         config.put(KafkaConsumerConfig.KAFKA_CONSUMER_CONFIG_PREFIX + ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
     }
 
-    private static final String BRIDGE_HOST = "127.0.0.1";
-    private static final int BRIDGE_PORT = 8080;
+    static final String BRIDGE_HOST = "127.0.0.1";
+    static final int BRIDGE_PORT = 8080;
 
-    // for periodic/multiple messages test
-    private static final int PERIODIC_MAX_MESSAGE = 10;
-    private static final int PERIODIC_DELAY = 200;
-    private static final int MULTIPLE_MAX_MESSAGE = 10;
-    private static final int TEST_TIMEOUT = 60;
-    private int count;
+
+    static final int TEST_TIMEOUT = 60;
+
 
     // for request configuration
     private static final long RESPONSE_TIMEOUT = 2000L;
 
-    private Vertx vertx;
+    Vertx vertx;
     private HttpBridge httpBridge;
-    private WebClient client;
+    WebClient client;
 
     private HttpBridgeConfig bridgeConfigProperties;
 
     //HTTP methods with configured Response timeout
-    private HttpRequest<Buffer> postRequest(String requestURI) {
+    HttpRequest<Buffer> postRequest(String requestURI) {
         return client.post(requestURI)
                 .timeout(RESPONSE_TIMEOUT);
     }
@@ -123,643 +114,6 @@ class HttpBridgeTest extends KafkaClusterTestBase {
     @AfterEach
     void after(VertxTestContext context) {
         vertx.close(context.succeeding(arg -> context.completeNow()));
-    }
-
-    @Test
-    void sendSimpleMessage(VertxTestContext context) throws Throwable {
-        String topic = "sendSimpleMessage";
-        kafkaCluster.createTopic(topic, 1, 1);
-
-        String value = "message-value";
-
-        JsonArray records = new JsonArray();
-        JsonObject json = new JsonObject();
-        json.put("value", value);
-        records.add(json);
-
-        JsonObject root = new JsonObject();
-        root.put("records", records);
-
-        postRequest("/topics/" + topic)
-                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
-                .putHeader("Content-Type", BridgeContentType.KAFKA_JSON_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(root, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-
-                        JsonArray offsets = bridgeResponse.getJsonArray("offsets");
-                        assertEquals(1, offsets.size());
-                        JsonObject metadata = offsets.getJsonObject(0);
-                        assertEquals(0, metadata.getInteger("partition"));
-                        assertEquals(0L, metadata.getLong("offset"));
-                    });
-                });
-
-        Properties config = kafkaCluster.useTo().getConsumerProperties("groupId", null, OffsetResetStrategy.EARLIEST);
-
-        KafkaConsumer<String, String> consumer = KafkaConsumer.create(this.vertx, config,
-                new StringDeserializer(), new KafkaJsonDeserializer<>(String.class));
-        consumer.handler(record -> {
-            context.verify(() -> {
-                assertEquals(value, record.value());
-                assertEquals(topic, record.topic());
-                assertEquals(0, record.partition());
-                assertEquals(0L, record.offset());
-                assertNull(record.key());
-            });
-            log.info("Message consumed topic={} partition={} offset={}, key={}, value={}",
-                    record.topic(), record.partition(), record.offset(), record.key(), record.value());
-            consumer.close();
-            context.completeNow();
-        });
-
-        consumer.subscribe(topic, done -> {
-            if (!done.succeeded()) {
-                context.failNow(done.cause());
-            }
-        });
-    }
-
-    @Test
-    void sendSimpleMessageToPartition(VertxTestContext context) {
-        String topic = "sendSimpleMessageToPartition";
-        kafkaCluster.createTopic(topic, 2, 1);
-
-        String value = "message-value";
-
-        int partition = 1;
-
-        JsonArray records = new JsonArray();
-        JsonObject json = new JsonObject();
-        json.put("value", value);
-        json.put("partition", partition);
-        records.add(json);
-
-        JsonObject root = new JsonObject();
-        root.put("records", records);
-
-        postRequest("/topics/" + topic)
-                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
-                .putHeader("Content-Type", BridgeContentType.KAFKA_JSON_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(root, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-
-                        JsonArray offsets = bridgeResponse.getJsonArray("offsets");
-                        assertEquals(1, offsets.size());
-                        JsonObject metadata = offsets.getJsonObject(0);
-                        assertEquals(partition, metadata.getInteger("partition"));
-                        assertEquals(0L, metadata.getLong("offset"));
-                    });
-                });
-
-        Properties config = kafkaCluster.useTo().getConsumerProperties("groupId", null, OffsetResetStrategy.EARLIEST);
-
-        KafkaConsumer<String, String> consumer = KafkaConsumer.create(this.vertx, config,
-                new StringDeserializer(), new KafkaJsonDeserializer<>(String.class));
-        consumer.handler(record -> {
-            context.verify(() -> {
-                assertEquals(value, record.value());
-                assertEquals(topic, record.topic());
-                assertEquals(partition, record.partition());
-                assertEquals(0L, record.offset());
-                assertNull(record.key());
-            });
-            log.info("Message consumed topic={} partition={} offset={}, key={}, value={}",
-                    record.topic(), record.partition(), record.offset(), record.key(), record.value());
-            consumer.close();
-            context.completeNow();
-        });
-
-        consumer.subscribe(topic, done -> {
-            if (!done.succeeded()) {
-                context.failNow(done.cause());
-            }
-        });
-    }
-
-    @Test
-    void sendSimpleMessageWithKey(VertxTestContext context) {
-        String topic = "sendSimpleMessageWithKey";
-        kafkaCluster.createTopic(topic, 2, 1);
-
-        String value = "message-value";
-        String key = "my-key";
-
-        JsonArray records = new JsonArray();
-        JsonObject json = new JsonObject();
-        json.put("value", value);
-        json.put("key", key);
-        records.add(json);
-
-        JsonObject root = new JsonObject();
-        root.put("records", records);
-
-        postRequest("/topics/" + topic)
-                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
-                .putHeader("Content-Type", BridgeContentType.KAFKA_JSON_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(root, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-
-                        JsonArray offsets = bridgeResponse.getJsonArray("offsets");
-                        assertEquals(1, offsets.size());
-                        JsonObject metadata = offsets.getJsonObject(0);
-                        assertNotNull(metadata.getInteger("partition"));
-                        assertEquals(0L, metadata.getLong("offset"));
-                    });
-                });
-
-        Properties config = kafkaCluster.useTo().getConsumerProperties("groupId", null, OffsetResetStrategy.EARLIEST);
-
-        KafkaConsumer<String, String> consumer = KafkaConsumer.create(this.vertx, config,
-                new KafkaJsonDeserializer<>(String.class), new KafkaJsonDeserializer<>(String.class));
-        consumer.handler(record -> {
-            context.verify(() -> {
-                assertEquals(value, record.value());
-                assertEquals(topic, record.topic());
-                assertNotNull(record.partition());
-                assertEquals(0L, record.offset());
-                assertEquals(record.key(), key);
-            });
-            log.info("Message consumed topic={} partition={} offset={}, key={}, value={}",
-                    record.topic(), record.partition(), record.offset(), record.key(), record.value());
-            consumer.close();
-            context.completeNow();
-        });
-
-        consumer.subscribe(topic, done -> {
-            if (!done.succeeded()) {
-                context.failNow(done.cause());
-            }
-        });
-    }
-
-    @Test
-    void sendBinaryMessageWithKey(VertxTestContext context) {
-        String topic = "sendBinaryMessageWithKey";
-        kafkaCluster.createTopic(topic, 2, 1);
-
-        String value = "message-value";
-        String key = "my-key";
-
-        JsonArray records = new JsonArray();
-        JsonObject json = new JsonObject();
-        json.put("value", DatatypeConverter.printBase64Binary(value.getBytes()));
-        json.put("key", DatatypeConverter.printBase64Binary(key.getBytes()));
-        records.add(json);
-
-        JsonObject root = new JsonObject();
-        root.put("records", records);
-
-        postRequest("/topics/" + topic)
-                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
-                .putHeader("Content-Type", BridgeContentType.KAFKA_JSON_BINARY)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(root, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-
-                        JsonArray offsets = bridgeResponse.getJsonArray("offsets");
-                        assertEquals(1, offsets.size());
-                        JsonObject metadata = offsets.getJsonObject(0);
-                        assertNotNull(metadata.getInteger("partition"));
-                        assertEquals(0L, metadata.getLong("offset"));
-                    });
-                });
-
-        Properties config = kafkaCluster.useTo().getConsumerProperties("groupId", null, OffsetResetStrategy.EARLIEST);
-
-        KafkaConsumer<byte[], byte[]> consumer = KafkaConsumer.create(this.vertx, config,
-                new ByteArrayDeserializer(), new ByteArrayDeserializer());
-        consumer.handler(record -> {
-            context.verify(() -> {
-                assertEquals(value, new String(record.value()));
-                assertEquals(topic, record.topic());
-                assertNotNull(record.partition());
-                assertEquals(0L, record.offset());
-                assertEquals(key, new String(record.key()));
-            });
-
-            log.info("Message consumed topic={} partition={} offset={}, key={}, value={}",
-                    record.topic(), record.partition(), record.offset(), record.key(), record.value());
-            consumer.close();
-            context.completeNow();
-        });
-
-        consumer.subscribe(topic, done -> {
-            if (!done.succeeded()) {
-                context.failNow(done.cause());
-            }
-        });
-    }
-
-    @Test
-    void sendPeriodicMessage(VertxTestContext context) {
-        String topic = "sendPeriodicMessage";
-        kafkaCluster.createTopic(topic, 1, 1);
-
-        Properties config = kafkaCluster.useTo().getConsumerProperties("groupId", null, OffsetResetStrategy.EARLIEST);
-
-        KafkaConsumer<String, String> consumer = KafkaConsumer.create(this.vertx, config,
-                new KafkaJsonDeserializer<>(String.class), new KafkaJsonDeserializer<>(String.class));
-
-        this.count = 0;
-
-        this.vertx.setPeriodic(HttpBridgeTest.PERIODIC_DELAY, timerId -> {
-
-            if (this.count < HttpBridgeTest.PERIODIC_MAX_MESSAGE) {
-
-                JsonArray records = new JsonArray();
-                JsonObject json = new JsonObject();
-                json.put("value", "Periodic message [" + this.count + "]");
-                json.put("key", "key-" + this.count);
-                records.add(json);
-
-                JsonObject root = new JsonObject();
-                root.put("records", records);
-
-                postRequest("/topics/" + topic)
-                        .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
-                        .putHeader("Content-Type", BridgeContentType.KAFKA_JSON_JSON)
-                        .sendJsonObject(root, ar -> { });
-
-                this.count++;
-            } else {
-                this.vertx.cancelTimer(timerId);
-
-                consumer.subscribe(topic, done -> {
-                    if (!done.succeeded()) {
-                        context.failNow(done.cause());
-                    }
-                });
-            }
-        });
-
-        consumer.batchHandler(records -> {
-            context.verify(() -> {
-                assertEquals(this.count, records.size());
-                for (int i = 0; i < records.size(); i++) {
-                    KafkaConsumerRecord<String, String> record = records.recordAt(i);
-                    log.info("Message consumed topic={} partition={} offset={}, key={}, value={}",
-                            record.topic(), record.partition(), record.offset(), record.key(), record.value());
-                    assertEquals(record.value(), "Periodic message [" + i + "]");
-                    assertEquals(topic, record.topic());
-                    assertNotNull(record.partition());
-                    assertNotNull(record.offset());
-                    assertEquals(record.key(), "key-" + i);
-                }
-            });
-
-            consumer.close();
-            context.completeNow();
-        });
-
-        consumer.handler(record -> { });
-    }
-
-    @Test
-    void sendMultipleMessages(VertxTestContext context) {
-        String topic = "sendMultipleMessages";
-        kafkaCluster.createTopic(topic, 1, 1);
-
-        String value = "message-value";
-
-        int numMessages = MULTIPLE_MAX_MESSAGE;
-
-        JsonArray records = new JsonArray();
-        for (int i = 0; i < numMessages; i++) {
-            JsonObject json = new JsonObject();
-            json.put("value", value + "-" + i);
-            records.add(json);
-        }
-        JsonObject root = new JsonObject();
-        root.put("records", records);
-
-        postRequest("/topics/" + topic)
-                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
-                .putHeader("Content-Type", BridgeContentType.KAFKA_JSON_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(root, ar -> {
-                    context.verify(() -> assertTrue(ar.succeeded()));
-
-                    HttpResponse<JsonObject> response = ar.result();
-                    assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                    JsonObject bridgeResponse = response.body();
-
-                    JsonArray offsets = bridgeResponse.getJsonArray("offsets");
-                    assertEquals(numMessages, offsets.size());
-                    for (int i = 0; i < numMessages; i++) {
-                        JsonObject metadata = offsets.getJsonObject(i);
-                        assertEquals(0, metadata.getInteger("partition"));
-                        assertNotNull(metadata.getLong("offset"));
-                    }
-                });
-
-        Properties config = kafkaCluster.useTo().getConsumerProperties("groupId", null, OffsetResetStrategy.EARLIEST);
-
-        KafkaConsumer<String, String> consumer = KafkaConsumer.create(this.vertx, config,
-                new StringDeserializer(), new KafkaJsonDeserializer<String>(String.class));
-        this.count = 0;
-        consumer.handler(record -> {
-            context.verify(() -> {
-                assertEquals(value + "-" + this.count++, record.value());
-                assertEquals(topic, record.topic());
-                assertNotNull(record.partition());
-                assertNotNull(record.offset());
-                assertNull(record.key());
-            });
-
-            log.info("Message consumed topic={} partition={} offset={}, key={}, value={}",
-                    record.topic(), record.partition(), record.offset(), record.key(), record.value());
-
-            if (this.count == numMessages) {
-                consumer.close();
-                context.completeNow();
-            }
-        });
-
-        consumer.subscribe(topic, done -> {
-            if (!done.succeeded()) {
-                context.failNow(done.cause());
-            }
-        });
-    }
-
-    @Test
-    void createConsumer(VertxTestContext context) throws InterruptedException {
-        String name = "my-kafka-consumer";
-        String groupId = "my-group";
-
-        String baseUri = "http://" + BRIDGE_HOST + ":" + BRIDGE_PORT + "/consumers/" + groupId + "/instances/" + name;
-
-        JsonObject json = new JsonObject();
-        json.put("name", name);
-        json.put("auto.offset.reset", "earliest");
-        json.put("enable.auto.commit", "true");
-        json.put("fetch.min.bytes", "100");
-
-        postRequest("/consumers/" + groupId)
-                .putHeader("Content-length", String.valueOf(json.toBuffer().length()))
-                .putHeader("Content-type", BridgeContentType.KAFKA_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(json, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-                        String consumerInstanceId = bridgeResponse.getString("instance_id");
-                        String consumerBaseUri = bridgeResponse.getString("base_uri");
-                        assertEquals(name, consumerInstanceId);
-                        assertEquals(baseUri, consumerBaseUri);
-                    });
-
-                    context.completeNow();
-                });
-
-        assertTrue(context.awaitCompletion(TEST_TIMEOUT, TimeUnit.SECONDS));
-    }
-
-    @Test
-    void createConsumerWithForwardedHeaders(VertxTestContext context) throws InterruptedException {
-        String name = "my-kafka-consumer";
-        String groupId = "my-group";
-
-        // this test emulates a create consumer request coming from an API gateway/proxy
-        String xForwardedHost = "my-api-gateway-host:443";
-        String xForwardedProto = "https";
-
-        String baseUri = xForwardedProto + "://" + xForwardedHost + "/consumers/" + groupId + "/instances/" + name;
-
-        JsonObject json = new JsonObject();
-        json.put("name", name);
-        json.put("auto.offset.reset", "earliest");
-        json.put("enable.auto.commit", "true");
-        json.put("fetch.min.bytes", "100");
-
-        postRequest("/consumers/" + groupId)
-                .putHeader("Content-length", String.valueOf(json.toBuffer().length()))
-                .putHeader("Content-type", BridgeContentType.KAFKA_JSON)
-                .putHeader("X-Forwarded-Host", xForwardedHost)
-                .putHeader("X-Forwarded-Proto", xForwardedProto)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(json, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-                        String consumerInstanceId = bridgeResponse.getString("instance_id");
-                        String consumerBaseUri = bridgeResponse.getString("base_uri");
-                        assertEquals(name, consumerInstanceId);
-                        assertEquals(baseUri, consumerBaseUri);
-                    });
-
-                    context.completeNow();
-                });
-
-        assertTrue(context.awaitCompletion(TEST_TIMEOUT, TimeUnit.SECONDS));
-    }
-
-    @Test
-    void createConsumerWithForwardedHeader(VertxTestContext context) throws InterruptedException {
-        String name = "my-kafka-consumer";
-        String groupId = "my-group";
-
-        // this test emulates a create consumer request coming from an API gateway/proxy
-        String forwarded = "host=my-api-gateway-host:443;proto=https";
-
-        String baseUri = "https://my-api-gateway-host:443/consumers/" + groupId + "/instances/" + name;
-
-        JsonObject json = new JsonObject();
-        json.put("name", name);
-        json.put("auto.offset.reset", "earliest");
-        json.put("enable.auto.commit", "true");
-        json.put("fetch.min.bytes", "100");
-
-        postRequest("/consumers/" + groupId)
-                .putHeader("Content-length", String.valueOf(json.toBuffer().length()))
-                .putHeader("Content-type", BridgeContentType.KAFKA_JSON)
-                .putHeader("Forwarded", forwarded)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(json, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-                        String consumerInstanceId = bridgeResponse.getString("instance_id");
-                        String consumerBaseUri = bridgeResponse.getString("base_uri");
-                        assertEquals(name, consumerInstanceId);
-                        assertEquals(baseUri, consumerBaseUri);
-                    });
-
-                    context.completeNow();
-                });
-
-        assertTrue(context.awaitCompletion(TEST_TIMEOUT, TimeUnit.SECONDS));
-    }
-
-    @Test
-    void createConsumerWithForwardedPathHeader(VertxTestContext context) throws InterruptedException {
-        String name = "my-kafka-consumer";
-        String groupId = "my-group";
-
-        // this test emulates a create consumer request coming from an API gateway/proxy
-        String forwarded = "host=my-api-gateway-host:443;proto=https";
-        String xForwardedPath = "/my-bridge/consumers/" + groupId;
-
-        String baseUri = "https://my-api-gateway-host:443/my-bridge/consumers/" + groupId + "/instances/" + name;
-
-        JsonObject json = new JsonObject();
-        json.put("name", name);
-        json.put("auto.offset.reset", "earliest");
-        json.put("enable.auto.commit", "true");
-        json.put("fetch.min.bytes", "100");
-
-        postRequest("/consumers/" + groupId)
-                .putHeader("Content-length", String.valueOf(json.toBuffer().length()))
-                .putHeader("Content-type", BridgeContentType.KAFKA_JSON)
-                .putHeader("Forwarded", forwarded)
-                .putHeader("X-Forwarded-Path", xForwardedPath)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(json, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-                        String consumerInstanceId = bridgeResponse.getString("instance_id");
-                        String consumerBaseUri = bridgeResponse.getString("base_uri");
-                        assertEquals(name, consumerInstanceId);
-                        assertEquals(baseUri, consumerBaseUri);
-                    });
-
-                    context.completeNow();
-                });
-
-        assertTrue(context.awaitCompletion(TEST_TIMEOUT, TimeUnit.SECONDS));
-    }
-
-    @Test
-    void createConsumerWithForwardedHeaderDefaultPort(VertxTestContext context) throws InterruptedException {
-        String name = "my-kafka-consumer";
-        String groupId = "my-group";
-
-        // this test emulates a create consumer request coming from an API gateway/proxy
-        String forwarded = "host=my-api-gateway-host;proto=http";
-
-        String baseUri = "http://my-api-gateway-host:80/consumers/" + groupId + "/instances/" + name;
-
-        JsonObject json = new JsonObject();
-        json.put("name", name);
-        json.put("auto.offset.reset", "earliest");
-        json.put("enable.auto.commit", "true");
-        json.put("fetch.min.bytes", "100");
-
-        postRequest("/consumers/" + groupId)
-                .putHeader("Content-length", String.valueOf(json.toBuffer().length()))
-                .putHeader("Content-type", BridgeContentType.KAFKA_JSON)
-                .putHeader("Forwarded", forwarded)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(json, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.OK.code(), response.statusCode());
-                        JsonObject bridgeResponse = response.body();
-                        String consumerInstanceId = bridgeResponse.getString("instance_id");
-                        String consumerBaseUri = bridgeResponse.getString("base_uri");
-                        assertEquals(name, consumerInstanceId);
-                        assertEquals(baseUri, consumerBaseUri);
-                    });
-
-                    context.completeNow();
-                });
-
-        assertTrue(context.awaitCompletion(TEST_TIMEOUT, TimeUnit.SECONDS));
-    }
-
-    @Test
-    void createConsumerWithForwardedHeaderWrongProto(VertxTestContext context) throws InterruptedException {
-        String name = "my-kafka-consumer";
-        String groupId = "my-group";
-
-        // this test emulates a create consumer request coming from an API gateway/proxy
-        String forwarded = "host=my-api-gateway-host;proto=mqtt";
-
-        JsonObject json = new JsonObject();
-        json.put("name", name);
-        json.put("auto.offset.reset", "earliest");
-        json.put("enable.auto.commit", "true");
-        json.put("fetch.min.bytes", "100");
-
-        postRequest("/consumers/" + groupId)
-                .putHeader("Content-length", String.valueOf(json.toBuffer().length()))
-                .putHeader("Content-type", BridgeContentType.KAFKA_JSON)
-                .putHeader("Forwarded", forwarded)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(json, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        HttpBridgeError error = HttpBridgeError.fromJson(response.body());
-                        assertEquals(HttpResponseStatus.INTERNAL_SERVER_ERROR.code(), response.statusCode());
-                        assertEquals(HttpResponseStatus.INTERNAL_SERVER_ERROR.code(), error.getCode());
-                        assertEquals("mqtt is not a valid schema/proto.", error.getMessage());
-                    });
-
-                    context.completeNow();
-                });
-
-        assertTrue(context.awaitCompletion(TEST_TIMEOUT, TimeUnit.SECONDS));
-    }
-
-    @Test
-    void createConsumerWithWrongParameter(VertxTestContext context) throws InterruptedException {
-        String name = "my-kafka-consumer";
-        String groupId = "my-group";
-
-        JsonObject json = new JsonObject();
-        json.put("name", name);
-        json.put("auto.offset.reset", "foo");
-
-        postRequest("/consumers/" + groupId)
-                .putHeader("Content-length", String.valueOf(json.toBuffer().length()))
-                .putHeader("Content-type", BridgeContentType.KAFKA_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(json, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.INTERNAL_SERVER_ERROR.code(), response.statusCode());
-                        HttpBridgeError error = HttpBridgeError.fromJson(response.body());
-                        assertEquals(HttpResponseStatus.INTERNAL_SERVER_ERROR.code(), error.getCode());
-                        assertEquals("Invalid value foo for configuration auto.offset.reset: String must be one of: latest, earliest, none",
-                                error.getMessage());
-                    });
-
-                    context.completeNow();
-                });
-
-        assertTrue(context.awaitCompletion(TEST_TIMEOUT, TimeUnit.SECONDS));
     }
 
     @Test
@@ -3238,72 +2592,9 @@ class HttpBridgeTest extends KafkaClusterTestBase {
         context.completeNow();
     }
 
-    @Test
-    void sendMessageLackingRequiredProperty(VertxTestContext context) throws Throwable {
-        String topic = "sendMessageLackingRequiredProperty";
-        kafkaCluster.createTopic(topic, 1, 1);
 
-        String key = "my-key";
 
-        JsonArray records = new JsonArray();
-        JsonObject json = new JsonObject();
-        json.put("key", key);
-        records.add(json);
 
-        JsonObject root = new JsonObject();
-        root.put("records", records);
-
-        postRequest("/topics/" + topic)
-                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
-                .putHeader("Content-Type", BridgeContentType.KAFKA_JSON_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(root, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        HttpBridgeError error = HttpBridgeError.fromJson(response.body());
-                        assertEquals(HttpResponseStatus.BAD_REQUEST.code(), response.statusCode());
-                        assertEquals(HttpResponseStatus.BAD_REQUEST.code(), error.getCode());
-                        assertEquals("Validation error on: body.records[0] - $.records[0].value: is missing but it is required",
-                                error.getMessage());
-                    });
-                    context.completeNow();
-                });
-    }
-
-    @Test
-    void sendMessageWithUnknownProperty(VertxTestContext context) throws Throwable {
-        String topic = "sendMessageWithUnknownProperty";
-        kafkaCluster.createTopic(topic, 1, 1);
-
-        String value = "message-value";
-
-        JsonArray records = new JsonArray();
-        JsonObject json = new JsonObject();
-        json.put("value", value);
-        json.put("foo", "unknown property");
-        records.add(json);
-
-        JsonObject root = new JsonObject();
-        root.put("records", records);
-
-        postRequest("/topics/" + topic)
-                .putHeader("Content-length", String.valueOf(root.toBuffer().length()))
-                .putHeader("Content-Type", BridgeContentType.KAFKA_JSON_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(root, ar -> {
-                    context.verify(() -> {
-                        assertTrue(ar.succeeded());
-                        HttpResponse<JsonObject> response = ar.result();
-                        assertEquals(HttpResponseStatus.BAD_REQUEST.code(), response.statusCode());
-                        HttpBridgeError error = HttpBridgeError.fromJson(response.body());
-                        assertEquals(HttpResponseStatus.BAD_REQUEST.code(), error.getCode());
-                        assertEquals("Validation error on: body.records[0] - $.records[0].foo: is not defined in the schema and the schema does not allow additional properties",
-                                error.getMessage());
-                    });
-                    context.completeNow();
-                });
-    }
 
     @Test
     void subscribeExclusiveTopicAndPattern(VertxTestContext context) throws Throwable {
